Extract utils
