+++ src/routes/event/index.tsx
@@ -12,7 +12,7 @@ import { mapMarker } from '@/icons/map-marker'
 import { googleMaps } from '@/icons/google-maps'
 import { infoOutline } from '@/icons/info-outline'
 import { video } from '@/icons/video'
-import { formatDateRange } from '@/utils'
+import { formatDateRange } from '@/utils/format-date-range'
 import Icon from '@/components/icon'
 import Chip from '@/components/chip'
 import style from './style.css'
