+++ src/utils/format-team-number/index.ts
@@ -1,0 +1,1 @@
+export const formatTeamNumber = (team: string) => team.replace(/^frc/, '')
