+++ src/utils/format-time/index.ts
@@ -1,0 +1,7 @@
+export const formatTime = (date: string, timeZone?: string) =>
+  new Date(date).toLocaleTimeString('en-US', {
+    hour: 'numeric',
+    minute: 'numeric',
+    weekday: 'short',
+    timeZone,
+  })
